+++ ruby/examples/webhook-server/main.rb
@@ -1,23 +1,14 @@
-require 'securerandom'
-require 'truelayer-signing'
-require 'webrick'
+require "http"
+require "truelayer-signing"
+require "webrick"
 
 class TrueLayerSigningExamples
   # Note: the webhook path can be whatever is configured for your application.
   # Here a unique path is used, matching the example signature in the README.
   WEBHOOK_PATH = "/hook/d7a2c49d-110a-4ed2-a07d-8fdb3ea6424b".freeze
-  PUBLIC_KEY_PEM = <<~TXT.freeze
-    -----BEGIN PUBLIC KEY-----
-    MIGbMBAGByqGSM49AgEGBSuBBAAjA4GGAAQBJ6ET9XeVCyMy+yOetZaNNCXPhwr5
-    BlyDDg1CLmyNM5SvqOs8RveL6dYl4lpPur4xrPQl04ggYlVd9wnHkZnp3jcBlXw8
-    Lc5phyYF1q2/QV/5wp2WHIhKDqUiXC0TvlE8d7MdTAN9yolcwrh6aWZ3kesTMZif
-    BgItyT6PXUab8mMdI8k=
-    -----END PUBLIC KEY-----
-  TXT
 
   class << self
     def run_webhook_server
-      TrueLayerSigning.certificate_id ||= SecureRandom.uuid
       server = WEBrick::HTTPServer.new(Port: 4567)
 
       puts "Server running at http://localhost:4567"
@@ -61,9 +52,19 @@ class TrueLayerSigningExamples
 
       return ["400", "Bad Request – Header `Tl-Signature` missing"] unless tl_signature
 
+      jku = TrueLayerSigning.extract_jws_header(tl_signature).jku
+
+      return ["400", "Bad Request – Signature missing `jku`"] unless jku
+      return ["401", "Unauthorized – Unpermitted `jku`"] \
+        unless jku == "https://webhooks.truelayer.com/.well-known/jwks" ||
+          jku == "https://webhooks.truelayer-sandbox.com/.well-known/jwks"
+
+      jwks = HTTP.get(jku)
+
+      return ["401", "Unauthorized – Unavailable `jwks` resource"] unless jwks.status.success?
+
       begin
-        TrueLayerSigning
-          .verify_with_pem(PUBLIC_KEY_PEM)
+        TrueLayerSigning.verify_with_jwks(jwks.to_s)
           .set_method(:post)
           .set_path(path)
           .set_headers(headers)