+++ ruby/lib/truelayer-signing.rb
@@ -28,8 +28,16 @@ module TrueLayerSigning
       Signer.new
     end
 
+    def extract_jws_header(signature)
+      Verifier.parse_tl_signature(signature).first
+    end
+
+    def verify_with_jwks(jwks)
+      Verifier.new(key_type: :jwks, key_value: jwks)
+    end
+
     def verify_with_pem(pem)
-      Verifier.new(key_value: pem)
+      Verifier.new(key_type: :pem, key_value: pem)
     end
   end
 end