+++ ruby/CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - ...
 
+## [0.2.0] – 2023-06-13
+
+- Add support for signature verification using a JWKS: `TrueLayerSigning.verify_with_jwks(jwks)`
+    and `TrueLayerSigning.extract_jws_header(signature)`.
+
 ## [0.1.2] – 2023-05-19
 
 - Fix conflict with JWT library